Use better way to implement
