+++ src/main/java/me/phyrextsai/hyperledger/crowdfunding/data/Campaign.java
@@ -27,7 +27,6 @@ public class Campaign {
         this.setOwner(owner);
         this.setInfo(info);
         this.setFundingAmount(fundingAmount);
-
     }
 
     public String getCampaignId() {
